+++ src/wiki/urls.py
@@ -2,7 +2,13 @@ from django.utils.module_loading import import_string
 from wiki.compat import include, url
 from wiki.conf import settings
 from wiki.core.plugins import registry
+from wiki import sites
 from wiki.views import accounts, article, deleted_list
+
+
+urlpatterns = [
+    url(r'^', sites.site.urls),
+]
 
 
 class WikiURLPatterns:
@@ -247,11 +253,20 @@ def get_pattern(app_name="wiki", namespace="wiki", url_config_class=None):
        single Django project.
        https://docs.djangoproject.com/en/dev/topics/http/urls/#topics-http-reversing-url-namespaces
     """
+    import warnings
+    warnings.warn(
+        "wiki.urls.get_pattern is deprecated and will be removed in next version, just `include('wiki.urls')` in your urlconf",
+        DeprecationWarning
+    )
     if url_config_class is None:
         url_config_classname = getattr(settings, 'URL_CONFIG_CLASS', None)
         if url_config_classname is None:
             url_config_class = WikiURLPatterns
         else:
+            warnings.warn(
+                "URL_CONFIG_CLASS is deprecated and will be removed in next version, override `wiki.sites.WikiSite` instead",
+                DeprecationWarning
+            )
             url_config_class = import_string(url_config_classname)
     urlpatterns = url_config_class().get_urls()